# apps/api/spec/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Execution Engine API
  version: 1.0.0
  description: Workflow definitions, versions, nodes, and their executions. See
    milestones.md for the full system this API is the first slice of.
components:
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          maxLength: 2000
        title:
          type: string
          maxLength: 255
        status:
          type: integer
          minimum: -2147483648
          maximum: 2147483647
          format: int32
        detail:
          type: string
          maxLength: 2000
        instance:
          type: string
          maxLength: 2000
        code:
          type: string
          maxLength: 100
        errors:
          maxItems: 100
          type: array
          items: {}
      required:
        - type
        - title
        - status
        - detail
        - instance
        - code
      additionalProperties: false
...
    WorkflowVersion:
      type: object
      properties:
        id:
          type: string
          maxLength: 36
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        workflowId:
          type: string
          maxLength: 36
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        version:
          type: integer
          minimum: 0
          exclusiveMinimum: true
          maximum: 2147483647
          format: int32
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
        createdAt:
          type: string
          maxLength: 35
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
        publishedAt:
          nullable: true
          type: string
          maxLength: 35
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
      required:
        - id
        - workflowId
        - version
        - status
        - createdAt
        - publishedAt
      additionalProperties: false
...
paths:
...
  /api/v1/workflows/{workflowId}/versions/{version}:
    get:
      operationId: getWorkflowVersion
      summary: Get a workflow version
      tags:
        - WorkflowVersions
      description: Returns a single workflow version by workflow ID and version number.
      parameters:
        - schema:
            type: string
            maxLength: 36
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: workflowId
          required: true
        - schema:
            type: string
            maxLength: 36
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: version
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
        "400":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      operationId: deleteWorkflowVersion
      summary: Delete a draft workflow version
      tags:
        - WorkflowVersions
      description: Deletes a draft version and its nodes. Published versions cannot be
        deleted.
      parameters:
        - schema:
            type: string
            maxLength: 36
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: workflowId
          required: true
        - schema:
            type: string
            maxLength: 36
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: version
          required: true
      responses:
        "204":
          description: Default Response
          content:
            application/json:
              schema:
                type: string
                nullable: true
                enum:
                  - null
        "400":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
servers:
  - url: http://localhost:3000
    description: Local development
    x-internal: true
...